requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
setuptools==59.6.0